@@ -2,8 +2,8 @@ class AccountInfo::MessagesController < AccountInfo::BaseController
   before_filter :load_target_user!
 
   def create
-    @message = MessagesManager.new(user: current_user.object).create(target_user: @target_user,
-                                                                     message: params[:message])
+    @message = MessagesManager.new(user: current_user.object).
+      create(target_user: @target_user, message: params[:message])
     json_append
   end
 
