@@ -4,7 +4,7 @@ class CommentsController < ApplicationController
   before_filter :load_comment!, only: [:edit, :update, :destroy, :make_visible, :hide]
 
   protect(:index, :create) { can? :see, @post }
-  protect(:edit, :update, :destroy) { can? :delete, @comment }
+  protect(:edit, :update, :make_visible, :hide, :destroy) { can? :manage, @comment }
 
   def index
     @query = Queries::Comments.new(post: @post, start_id: params[:last_comment_id])
@@ -12,8 +12,7 @@ class CommentsController < ApplicationController
   end
 
   def create
-    @comment = CommentManager.new(user: current_user.object, post: @post).create(message: params[:message],
-                                                                                 mentions: params[:mentions])
+    @comment = CommentManager.new(user: current_user.object, post: @post).create(params.slice(:message, :mentions))
     json_render
   end
 
@@ -23,17 +22,17 @@ class CommentsController < ApplicationController
 
   def update
     @comment.update_attributes(message: params[:message])
-    json_replace html: render_to_string(partial: 'comment_row', locals: {comment: @comment})
+    render_comment_row
   end
 
   def make_visible
     CommentManager.new(user: current_user.object, comment: @comment).show
-    json_replace html: render_to_string(partial: 'comment_row', locals: {comment: @comment})
+    render_comment_row
   end
 
   def hide
     CommentManager.new(user: current_user.object, comment: @comment).hide
-    json_replace html: render_to_string(partial: 'comment_row', locals: {comment: @comment})
+    render_comment_row
   end
 
   def destroy
@@ -50,4 +49,9 @@ class CommentsController < ApplicationController
   def load_post!
     @post = Post.where(id: params[:post_id]).first or error(404)
   end
+
+  def render_comment_row
+    json_replace partial: 'comment_row', locals: {comment: @comment}
+  end
 end
+
