@@ -2,16 +2,13 @@ class DocumentPostsController < PendingPostsController
 
   def cancel
     PostManager.new(user: current_user.object).cancel_pending_documents
-    json_render html: render_to_string('new', layout: false)
+    json_render template: 'new'
   end
 
   protected
 
   def create_post
-    PostManager.new(user: current_user.object).create_document_post title:         params[:title],
-                                                                    keywords_text: params[:keywords_text],
-                                                                    notify:        params[:notify],
-                                                                    message:       params[:message]
+    PostManager.new(user: current_user.object).create_document_post(params.slice(%i(title keyword_text message notify)))
   end
 
   def cancel_media_posts_path
@@ -23,3 +20,4 @@ class DocumentPostsController < PendingPostsController
   end
   helper_method :media_posts_path, :cancel_media_posts_path
 end
+
