@@ -1,8 +1,10 @@
 class MessagesController < ApplicationController
   before_filter :load_target_user!
 
+  protect { can? :send_message_to, @target_user }
+
   def new
-    json_success popup: render_to_string('new', layout: false)
+    json_popup
   end
 
   def create
@@ -17,3 +19,4 @@ class MessagesController < ApplicationController
     @target_user = User.where(id: params[:user_id]).first or error(404)
   end
 end
+
