@@ -6,11 +6,9 @@ class Owner::SecondStepsController < Owner::BaseController
   end
 
   def update
-    UserProfileManager.new(@user).update cost:           params[:cost],
-                                         profile_name:   params[:profile_name],
-                                         holder_name:    params[:holder_name],
-                                         routing_number: params[:routing_number],
-                                         account_number: params[:account_number]
+    UserProfileManager.new(@user).
+      update(params.slice(:cost, :profile_name, :holder_name, :routing_number, :account_number))
+
     if @user.has_full_account?
       notice(:congrats)
       json_redirect profile_path(@user.reload.slug)
