@@ -29,12 +29,12 @@ class PhotosController < UploadsController
 
   def create
     manager.create_pending_photos(params[:transloadit])
-    json_replace html: render_to_string(partial: 'photo_posts/pending_uploads')
+    json_replace partial: 'photo_posts/pending_uploads'
   end
 
   def destroy
     super
-    json_render html: render_to_string(partial: 'photo_posts/pending_uploads')
+    json_render partial: 'photo_posts/pending_uploads'
   end
 
   private
