@@ -9,12 +9,18 @@ class ProfileTypesController < ApplicationController
 
   def create
     UserProfileManager.new(current_user.object).add_profile_type(params['type'])
-    json_replace html: render_to_string(partial: 'list_assigned', locals: {profile_types: current_user.profile_types})
+    json_replace html: types_html
   end
 
   def destroy
     profile_type = ProfileType.where(id: params[:id]).first or error(404)
     UserProfileManager.new(current_user.object).remove_profile_type(profile_type)
-    json_replace html: render_to_string(partial: 'list_assigned', locals: {profile_types: current_user.profile_types})
+    json_replace html: types_html
+  end
+
+  private
+
+  def types_html
+    render_to_string(partial: 'list_assigned', locals: {profile_types: current_user.profile_types})
   end
 end
