@@ -3,15 +3,12 @@ class RssFeedsController < ApplicationController
 
   def show
     headers['Content-Type'] = 'application/xml; charset=utf-8'
-    @feed_events = FeedEvent.where(subscription_target_user_id: @user.id).order('feed_events.created_at DESC').limit(70).to_a
+    @feed_events = FeedEvent.where(subscription_target_user_id: @user.id).
+      order('feed_events.created_at DESC').limit(70).to_a
 
     respond_to do |format|
-      format.atom do
-        render layout: false
-      end
-      format.rss do
-        render layout: false
-      end
+      format.atom { render layout: false }
+      format.rss { render layout: false }
     end
   end
 
