@@ -2,16 +2,13 @@ class VideoPostsController < MediaPostsController
 
   def cancel
     PostManager.new(user: current_user.object).cancel_pending_videos
-    json_render html: render_to_string('new', layout: false)
+    json_render template: 'new'
   end
 
   protected
 
   def create_post
-    PostManager.new(user: current_user.object).create_video_post title:         params[:title],
-                                                                 keywords_text: params[:keywords_text],
-                                                                 notify:        params[:notify],
-                                                                 message:       params[:message]
+    PostManager.new(user: current_user.object).create_video_post(params.slice(%i(title keyword_text message notify)))
   end
 
   def cancel_media_posts_path
