@@ -2,14 +2,14 @@ class MessagesManager < BaseManager
   attr_reader :user
 
   # @param user [User] Who is sending the message
-  def initialize(user: user)
+  def initialize(user: )
     @user = user
   end
 
   # @param target_user [User]
   # @param message [String]
   # @return [Message]
-  def create(target_user: target_user, message: text)
+  def create(target_user: , message: )
     fail_with! message: :empty if message.blank?
     fail_with! message: :too_long if message.length > 1000
 
