@@ -18,15 +18,14 @@ class SubscriptionManager < BaseManager
   # @param expiry_year [String]
   # @param expiry_month [String]
   # @return [Subscription]
-  def register_subscribe_and_pay email: nil,
+  def register_subscribe_and_pay(email: nil,
                                  full_name: nil,
                                  password: nil,
                                  number: nil,
                                  cvc: nil,
                                  expiry_month: nil,
                                  expiry_year: nil,
-                                target: (raise ArgumentError)
-
+                                 target: )
     unless target.is_a?(Concerns::Subscribable)
       raise ArgumentError, "Cannot subscribe to #{target.class.name}"
     end
@@ -69,12 +68,11 @@ class SubscriptionManager < BaseManager
   # @param expiry_year [String]
   # @param expiry_month [String]
   # @return [Subscription]
-  def update_cc_subscribe_and_pay number: nil,
+  def update_cc_subscribe_and_pay(number: nil,
                                   cvc: nil,
                                   expiry_month: nil,
                                   expiry_year: nil,
-                                  target: (raise ArgumentError)
-
+                                  target: )
     unless target.is_a?(Concerns::Subscribable)
       raise ArgumentError, "Cannot subscribe to #{target.class.name}"
     end
@@ -135,8 +133,8 @@ class SubscriptionManager < BaseManager
     end
   end
 
-  def restore(subscription)
-    PaymentManager.new.pay_for(subscription, 'Payment for subscription') unless subscription.paid?
+  def restore(subscription) # TODO (DJ): move to constructor
+    PaymentManager.new.pay_for!(subscription, 'Payment for subscription') unless subscription.paid?
     subscription.restore!
 
     target_user = subscription.target_user
@@ -145,21 +143,38 @@ class SubscriptionManager < BaseManager
   end
 
   # @param subscription [Subscription]
-  def unsubscribe(subscription)
+  def unsubscribe(subscription) # TODO (DJ): move to constructor
     subscription.remove!
 
     target_user = subscription.target_user
     UserStatsManager.new(target_user).log_subscriptions_count
-    UnsubscribedFeedEvent.create! target_user: target_user, target: @subscriber
+
+    if subscription.rejected?
+      # TODO: create another type of event
+    else
+      UnsubscribedFeedEvent.create! target_user: target_user, target: @subscriber
+    end
   end
 
-  def enable_notifications(subscription)
+  def enable_notifications(subscription) # TODO (DJ): move to constructor
     subscription.notifications_enabled = true
     save_or_die! subscription
   end
 
-  def disable_notifications(subscription)
+  def disable_notifications(subscription) # TODO (DJ): move to constructor
     subscription.notifications_enabled = false
+    save_or_die! subscription
+  end
+
+  def reject(subscription) # TODO (DJ): move to constructor
+    subscription.rejected = true
+    subscription.rejected_at = Time.zone.now if subscription.rejected_at.nil?
+    save_or_die! subscription
+  end
+
+  def accept(subscription) # TODO (DJ): move to constructor
+    subscription.rejected = false
+    subscription.rejected_at = nil
     save_or_die! subscription
   end
 end
