@@ -1,10 +1,10 @@
 require 'spec_helper'
 
-describe Admin::StaffsController do
+describe Admin::StaffsController, type: :controller do
   before { sign_in create_admin }
 
   describe 'GET #index' do
     subject { get 'index' }
-    its(:status) { should == 200 }
+    it { should be_success }
   end
 end
