@@ -1,21 +1,7 @@
 require 'spec_helper'
 
-describe AudioPostsController do
+describe AudioPostsController, type: :controller do
   let(:owner) { create_user email: 'owner@gmail.com', is_profile_owner: true }
-
-  describe 'DELETE #cancel' do
-    subject { delete :cancel }
-
-    context 'unauthorized access' do
-      its(:status) { should == 401 }
-    end
-
-    context 'authorized access' do
-      before { sign_in owner }
-      its(:status) { should == 200 }
-      its(:body) { should match_regex /success/ }
-    end
-  end
 
   describe 'GET #new' do
     subject { get :new }
@@ -26,18 +12,32 @@ describe AudioPostsController do
 
     context 'authorized access' do
       before { sign_in owner }
-      its(:status) { should == 200 }
+      it { should be_success }
+    end
+  end
+
+  describe 'DELETE #cancel' do
+    subject { delete :cancel }
+
+    context 'unauthorized access' do
+      its(:status) { should == 401 }
+    end
+
+    context 'authorized access' do
+      before { sign_in owner }
+      it { should be_success }
+      its(:body) { should match_regex /success/ }
     end
   end
 
   describe 'POST #create' do
-    subject { post :create, title: 'aa', message: 'bb'}
+    subject { post :create, title: 'audio', message: 'post'}
 
     context 'authorized access' do
       before { sign_in owner }
-      let!(:pending_audio) { create_audios_upload(owner).first }
+      let!(:pending_audio) { create_audio_upload(owner).first }
 
-      its(:status) { should == 200 }
+      it { should be_success }
       its(:body) { should match_regex /replace/ }
     end
 
@@ -46,3 +46,4 @@ describe AudioPostsController do
     end
   end
 end
+
