@@ -1,9 +1,10 @@
 require 'spec_helper'
 
-describe CommentsController do
+describe CommentsController, type: :controller do
   let(:poster) { create_user email: 'poster@gmail.com' }
   let(:commenter) { create_user email: 'commenter@gmail.com' }
   let(:_post) { PostManager.new(user: poster).create_status_post(message: 'some post') }
+  let(:comment) { CommentManager.new(user: commenter, post: _post).create(message: 'test') }
 
   describe 'GET #index' do
     subject { get 'index', post_id: _post.id }
@@ -16,13 +17,13 @@ describe CommentsController do
     context 'authorized access' do
       context 'as poster' do
         before { sign_in poster }
-        its(:status) { should == 200 }
+        it { should be_success }
       end
 
       context 'as subscriber' do
         before { SubscriptionManager.new(commenter).subscribe_to(poster) }
         before { sign_in commenter }
-        its(:status) { should == 200 }
+        it { should be_success }
       end
     end
   end
@@ -38,19 +39,143 @@ describe CommentsController do
     context 'authorized access' do
       context 'as poster' do
         before { sign_in poster }
-        its(:status) { should == 200 }
+        it { should be_success }
       end
 
       context 'as subscriber' do
         before { SubscriptionManager.new(commenter).subscribe_to(poster) }
         before { sign_in commenter }
+        it { should be_success }
+      end
+    end
+  end
+
+  describe 'GET #edit' do
+    before { SubscriptionManager.new(commenter).subscribe_to(poster) }
+    subject { get 'edit', post_id: _post.id, id: comment.id }
+
+    context 'unauthorized access' do
+      its(:status) { should == 401 }
+    end
+
+    context 'authorized access' do
+      context 'as comment owner' do
+        before { sign_in commenter }
+        it { should be_success }
+      end
+
+      context 'as a post owner' do
+        before { sign_in poster }
         its(:status) { should == 200 }
+      end
+
+      context 'as anybody else' do
+        before { sign_in }
+        its(:status) { should == 401 }
+      end
+    end
+  end
+
+  describe 'PUT #update' do
+    before { SubscriptionManager.new(commenter).subscribe_to(poster) }
+    subject(:perform_request) { put 'update', post_id: _post.id, id: comment.id, message: 'updated' }
+
+    context 'unauthorized access' do
+      its(:status) { should == 401 }
+    end
+
+    context 'authorized access' do
+      context 'as comment owner' do
+        before { sign_in commenter }
+        it { should be_success }
+
+        specify do
+          expect { perform_request }.to change { comment.reload.message }.to('updated')
+        end
+      end
+
+      context 'as a post owner' do
+        before { sign_in poster }
+        its(:status) { should == 200 }
+      end
+
+      context 'as anybody else' do
+        before { sign_in }
+        its(:status) { should == 401 }
+      end
+    end
+  end
+
+  describe 'PUT #make_visible' do
+    before do
+      SubscriptionManager.new(commenter).subscribe_to(poster)
+      CommentManager.new(comment: comment).hide
+    end
+
+    subject(:perform_request) { put 'make_visible', post_id: _post.id, id: comment.id }
+
+    context 'unauthorized access' do
+      its(:status) { should == 401 }
+    end
+
+    context 'authorized access' do
+      context 'as comment owner' do
+        before { sign_in commenter }
+        it { should be_success }
+
+        specify do
+          expect { perform_request }.to change { comment.reload.hidden? }.to(false)
+        end
+      end
+
+      context 'as a post owner' do
+        before { sign_in poster }
+        its(:status) { should == 200 }
+      end
+
+      context 'as anybody else' do
+        before { sign_in }
+        its(:status) { should == 401 }
+      end
+    end
+  end
+
+  describe 'PUT #hide' do
+    before do
+      SubscriptionManager.new(commenter).subscribe_to(poster)
+    end
+
+    subject(:perform_request) { put 'hide', post_id: _post.id, id: comment.id }
+
+    context 'unauthorized access' do
+      its(:status) { should == 401 }
+    end
+
+    context 'authorized access' do
+      context 'as comment owner' do
+        before { sign_in commenter }
+        it { should be_success }
+
+        specify do
+          expect { perform_request }.to change { comment.reload.hidden? }.to(true)
+        end
+      end
+
+      context 'as a post owner' do
+        before { sign_in poster }
+        its(:status) { should == 200 }
+      end
+
+      context 'as anybody else' do
+        before { sign_in }
+        its(:status) { should == 401 }
       end
     end
   end
 
   describe 'DELETE #destroy' do
-    let(:comment){ _post.comments.create(message: 'hello', user_id: commenter.id) }
+    before { SubscriptionManager.new(commenter).subscribe_to(poster) }
+    let(:comment) { CommentManager.new(user: commenter, post: _post).create(message: 'test') }
     subject { delete 'destroy', post_id: _post.id, id: comment.id }
 
     context 'unauthorized access' do
@@ -60,13 +185,19 @@ describe CommentsController do
     context 'authorized access' do
       context 'as comment owner' do
         before { sign_in commenter }
+        it { should be_success }
+      end
+
+      context 'as a post owner' do
+        before { sign_in poster }
         its(:status) { should == 200 }
       end
 
-      context 'as not comment owner' do
-        before { sign_in poster }
+      context 'as anybody else' do
+        before { sign_in }
         its(:status) { should == 401 }
       end
     end
   end
 end
+
