@@ -1,21 +1,7 @@
 require 'spec_helper'
 
-describe DocumentPostsController do
+describe DocumentPostsController, type: :controller do
   let(:owner) { create_user email: 'owner@gmail.com', is_profile_owner: true }
-
-  describe 'DELETE #cancel' do
-    subject { delete :cancel }
-
-    context 'unauthorized access' do
-      its(:status) { should == 401 }
-    end
-
-    context 'authorized access' do
-      before { sign_in owner }
-      its(:status) { should == 200 }
-      its(:body) { should match_regex /success/ }
-    end
-  end
 
   describe 'GET #new' do
     subject { get :new }
@@ -26,7 +12,21 @@ describe DocumentPostsController do
 
     context 'authorized access' do
       before { sign_in owner }
-      its(:status) { should == 200 }
+      it { should be_success }
+    end
+  end
+
+  describe 'DELETE #cancel' do
+    subject { delete :cancel }
+
+    context 'unauthorized access' do
+      its(:status) { should == 401 }
+    end
+
+    context 'authorized access' do
+      before { sign_in owner }
+      it { should be_success }
+      its(:body) { should match_regex /success/ }
     end
   end
 
@@ -35,9 +35,9 @@ describe DocumentPostsController do
 
     context 'authorized access' do
       before { sign_in owner }
-      let!(:pending_document) { create_documents_upload(owner).first }
+      let!(:pending_document) { create_document_upload(owner).first }
 
-      its(:status) { should == 200 }
+      it { should be_success }
       its(:body) { should match_regex /replace/ }
     end
 
