@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe DocumentsController do
+describe DocumentsController, type: :controller do
   let(:owner) { create_user email: 'owner@gmail.com', is_profile_owner: true }
 
   describe 'POST #create' do
@@ -12,12 +12,12 @@ describe DocumentsController do
 
     context 'authorized access' do
       before { sign_in owner }
-      its(:status) { should == 200 }
+      it { should be_success }
     end
   end
 
   describe 'DELETE #destroy' do
-    let(:document_upload) { create_documents_upload(owner).first  }
+    let(:document_upload) { create_document_upload(owner).first  }
     subject { delete 'destroy', id: document_upload.id }
 
     context 'unauthorized access' do
@@ -26,7 +26,7 @@ describe DocumentsController do
 
     context 'authorized access' do
       before { sign_in owner }
-      its(:status) { should == 200 }
+      it { should be_success }
     end
   end
 end
