@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe SubscriptionsController do
+describe SubscriptionsController, type: :controller do
   let(:owner) {
     create_user.tap do |user|
       UserProfileManager.new(user).create_profile_page
@@ -18,13 +18,13 @@ describe SubscriptionsController do
 
     context 'authorized access' do
       before { sign_in }
-      its(:status) { should == 200 }
+      it { should be_success }
     end
   end
 
   describe 'GET #new' do
     subject { get 'new', user_id: owner.slug }
-    its(:status) { should == 200 }
+    it { should be_success }
   end
 
   describe 'POST #create' do
@@ -45,7 +45,7 @@ describe SubscriptionsController do
 
       before { sign_in subscriber }
 
-      its(:status) { should == 200 }
+      it { should be_success }
     end
   end
 
