@@ -8,7 +8,7 @@ describe Queries::ProfileTypes do
   let!(:not_mathing_profile_type) { ProfileTypeManager.new.create(title: 'something really different') }
   let!(:user_matching_profile_type) do
     ProfileTypeManager.new.create(title: 'matching user').tap do |profile_type|
-      UserProfileManager.new(user).add_profile_type(profile_type)
+      UserProfileManager.new(user).add_profile_type(profile_type.title)
     end
   end
 
