@@ -1,13 +1,13 @@
 require 'spec_helper'
 
 describe Queries::Users do
+  before { create_user(email: 'spopov@gmail.com', full_name: 'Slava') }
+
   let!(:first_user) { create_user(email: 'szinin@gmail.com', full_name: 'Sergei') }
-  let!(:second_user) { create_user(email: 'spopov@gmail.com', full_name: 'Slava') }
-  let!(:performer) { create_user(email: 'performer@gmail.com') }
-  let!(:third_user) { create_user(email: 'third@gmail.com') }
+  let(:performer) { create_user(email: 'performer@gmail.com') }
 
-  subject { described_class.new(user: performer, query: 'szinin@gmail.com, spopov@gmail.com, performer@gmail.com') }
+  subject { described_class.new(user: performer, query: 'szinin@gmail.com') }
 
-  its(:results) { should =~ [first_user, second_user] }
-  its(:emails) { should =~ ['szinin@gmail.com', 'spopov@gmail.com'] }
+  specify { expect(subject.results).to eql [first_user] }
+  specify { expect(subject.emails).to eql ['szinin@gmail.com'] }
 end
